fixed Modal, again
