+++ src/components/ModalOverlay/ModalOverlay.module.css
@@ -1,15 +1,16 @@
-.overlay {
+.popup {
     display: flex;
-    position: fixed;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgba(0,0,0, 0.5);
-    justify-content: center;
-    align-items: center;
-    top: 0;
-    left: 0;
+    flex-direction: column;
+    min-width: 500px;
+    min-height: 600px;
+    background-color: ;
+    background-color: #1C1C21;
+    border-radius: 40px;
+    padding: 40px;
 }
 
-.overlayHidden {
-    display: none;
+.container {
+    display: flex;
+    justify-content: space-between;
+    width: calc(100% - 40px);
 }
