+++ src/components/app/BurgerIngredients/ModalInfoIngredients/IngredientDetails.jsx
@@ -2,7 +2,7 @@ import React, {useEffect} from 'react';
 
 import styles from './IngredientDetails.module.css';
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components';
-import { CLOSE_INFO_MODAL_WINDOW } from '../../../services/actions/currentIngredientsToModalAction';
+import { CLOSE_INFO_MODAL_WINDOW } from '../../../../services/actions/currentIngredientsToModalAction';
 import InfoOfIngridient from './InfoOfIngredient/InfoOfIngredinet';
 import { useSelector, useDispatch } from 'react-redux';
