+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./components/app/app";
 import reportWebVitals from "./reportWebVitals";
-import { rootReducer } from "./components/services/reducers/indexReducer";
+import { rootReducer } from "./services/reducers/indexReducer";
 import { applyMiddleware, createStore } from "redux";
 import { Provider } from "react-redux";
 import thunk from "redux-thunk";
