+++ src/services/actions/orderedIngredientsAction.jsx
@@ -1,5 +1,5 @@
-import { baseUrl } from "../../../utils/constants";
-import { checkResponse } from "../../../utils/constants";
+import { baseUrl } from "../../utils/constants";
+import { checkResponse } from "../../utils/constants";
 export const COLLECT_ORDER = 'COLLECT_ORDER';
 export const OPEN_MODAL_WINDOW = 'OPEN_MODAL_WINDOW';
 export const CLOSE_MODAL_WINDOW = 'CLOSE_MODAL_WINDOW';
